requirements.txt
# Core dependencies for the robot vision comparison project
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
scikit-image>=0.19.0
Pillow>=9.0.0
matplotlib>=3.5.0
tqdm>=4.62.0
joblib>=1.1.0
